qi/ki: full GOOS=plan9 lib_core/libc support, running hello_unix.c end to end
Adds svc_power.s/svc_sparc.s (raw Plan9 syscall stubs, transcribed
from the real Plan9 4th edition 9syscall/mkfile generator and verified
against qc -S/kc -S) plus the matching syscall_plan9_{power,sparc}
decl/h/zsyscall files, giving both archs full lib_core/libc GOOS=plan9
support at arm/mips parity.

Building hello_unix.c against it surfaced four more real bugs beyond
the raw-syscall level:

- qi: bx()/bcx() had the same missing-sign-extension bug as 5i's
  historical Ib()/Ibl() -- backward branches computed a huge positive
  offset instead of negative on this 64-bit host.
- ki: call() had a far more severe version of the same bug, breaking
  every function call (not just backward ones) by leaking CALL's own
  opcode bit into bit 32+ of the computed target.
- qi: setcr()/cmp()/cmpi()/twi()/tw() copied reg.r[] (u32int) into
  `long` locals, zero-extending instead of sign-extending negative
  results -- found via __fmtpad looping until timeout on a negative
  "remaining width" read back as huge-positive.
- ki: SPARC v8's umul/smul/udiv/sdiv were never implemented at all
  (op2[] carried bare `undef` at their slots), even though
  mkfiles/sparc/mkfile's own -M flag already assumed kl could emit
  them directly.

Also adds arch/sparc/rt0_plan9.s: arch/sparc/rt0.s is genuinely
Linux-specific (works around a register-window save-area shift only
the raw Linux SPARC32 kernel entry has), incompatible with Plan9's own
simpler process-entry convention -- selected via the same RT0OFILE
override mechanism arch/amd64,arm64/rt0_darwin.s already use.

demos/mkfile's demo_plan9 target now runs hello_unix.c to full
completion under qi and ki, matching the existing arm/5i and mips/vi
entries. Full top-level `mk test` passes clean throughout.

+++ machines/qi/branch.c
@@ -138,7 +138,11 @@ bcx(ulong ir)
 {
 	int bo, bi, xx;
 	ulong ea;
-	long imm;
+	// claude: int32, not long -- same fix, same reason as bx()'s
+	// identical comment above ("imm |= 0xFFFF0000" needs to happen at
+	// real 32-bit width, or a negative backward-branch offset comes
+	// out positive once widened to this host's 64-bit long).
+	int32 imm;
 	static char *opc[] = {"bc", "bcl", "bca", "bcla"};
 
 	getbobi(ir);
@@ -239,7 +243,22 @@ void
 bx(ulong ir)
 {
 	ulong ea;
-	long imm;
+	// claude: int32, not long -- `imm` was a `long` (64 bits on this
+	// host), and "imm |= 0xFC000000" only sets bits 26..31, leaving
+	// bits 32..63 zero: a real *negative* (backward) branch offset
+	// came out as a huge *positive* 64-bit number instead of sign-
+	// extending, so "ea = reg.pc + imm" below landed nowhere near the
+	// real target. Declaring `imm` as int32 directly makes the OR
+	// happen at the correct 32-bit width (the compiler reinterprets
+	// the bit pattern as signed there), and the later `reg.pc + imm`
+	// then sign-extends `imm` correctly when widening it to ulong --
+	// same class of bug already fixed in machines/5i/run.c's Ib()/
+	// Ibl() (see docs/claude_notes/notes_abi_plan9.txt's "A genuine,
+	// previously-undiscovered bug in 5i itself" section). Found via a
+	// real hello_unix.c (GOOS=plan9 lib_core/libc) segfaulting with a
+	// wild PC, despite the raw-syscall-only hello_plan9_power.s test
+	// (straight-line code, no backward branches at all) running fine.
+	int32 imm;
 	static char *opc[] = {"b", "bl", "ba", "bla"};
 
 	imm = ir & 0x03FFFFFC;
